Fixing bugs in CyclicInteger Writting some comments and type annotation to make the code more understandable and easier to work with, numpy arrays are still a PITA to type ;)
